Restructure debugging framework and use Cursive instead of tui-rs
This removes the `Debugger` trait from `mahboi::env`. I think this can be
done way better in another way. In particular, logging events can be done
via the `log` crate. This also means that `Emulator` doesn't store a
debugger anymore.

Additionally, the tui-rs debugger was replaced by a Cursive based one. It
is not implemented yet.