Move event handling to main thread
This was done because the previous approach does not work for MacOS, as
there, interactions with the window have to happen in the main thread.
Luckily, changing this wasn't a huge problem. Better even, I like the
new structure a bit better and it makes a few things easier. Does it
work on MacOS now? No idea, it's difficult for me to test. But it
might :P