Remove unfinished web version
This version is not even close to done. It is just a rainbow in a
canvas. I would work on that, I would probably redo at least half of it
as the wasm ecosystem changed anyway. And if no one uses it anyway, it
is just annoying in the repo. All the data is in the git history, so
nothing is lost.